pyproject.toml
[project]
name = "bbbbb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.0",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "torch>=2.9.1",
    "transformers>=4.57.1",
]
